feature(frontend): Add translation tool for multi language support in the app.
The `localizator`-function was added to provide support for multiple
languages.

It is provided globally via the `provide` method of the app.

The translator will receive a key and optional params which are replaced
when providing args.

When no matching value for a key is found in the current navigator
language the default language will be searched for a key. If there is no
translation found too, the key will be returned instead